requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PureIO==0.2.1
agate==1.6.1
agate-dbf==0.2.0
agate-excel==0.2.2
agate-sql==0.5.3
Babel==2.5.3
cycler==0.10.0
dbfread==2.0.7
et-xmlfile==1.0.1
future==0.16.0
isodate==0.6.0
jdcal==1.3
kiwisolver==1.0.1
leather==0.3.3
matplotlib==2.2.2
mock==2.0.0
numpy==1.14.2
openpyxl==2.5.1
pandas==0.22.0
parsedatetime==2.4
pbr==4.0.2
pi-ina219==1.1.0
pkg-resources==0.0.0
psutil==5.4.5
pyparsing==2.2.0
pyserial==3.4
python-dateutil==2.7.2
python-slugify==1.2.5
pytimeparse==1.1.7
pytz==2018.3
six==1.11.0
spidev==3.2
SQLAlchemy==1.2.6
Unidecode==1.0.22
xlrd==1.1.0
